Adjust the test suite to have a separate pytest style tests directory
