+++ tests/unit/conftest.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+
+from tests.support.pytest.fixtures import *  # pylint: disable=unused-wildcard-import
